Bug 1057341 - The log viewer overview should display properties from the job
This gets the job object and populates the "fields" section of the log viewer from that, rather than the ``header``
section of the ``text_log_summary`` artifact.  This way, Task Cluster jobs don't have to jump through hoops to add
it to the log summary.

This also fixes some coloring to match what we use in the main Treeherder view for result values.